Updated Animations
Background object and character animations with laying idle, flip up, standing idle and flip down for both types.
Created a secondary type of flipper for characters so they can be flipped separately from the background objects.
Working on adding prefabs of each object that will be pushed in the next commit.
Fixed the bug that caused jittering in background objects animation transitions.