Testing Two Enemies
Created a copy of the VS_Playtesting scene titled “TwoEnemyTest” to see if everything would work if a second enemy was added.
I tested this scene multiple times and was able to both win and lose without any weird enemy issues occurring. The enemies seem to be functioning properly on my end, however someone else should definitely test as well to see if they encounter any weird bugs that happen.