TileBehaviour Highlight and naming changes
Tile Behaviour now has 2 new functions, one that sets the color of the highlight and another that either activates or deactivates the highlight. updates were made to function calls in playerBehaviour and button manager in order to work with the new highlighting tile functions. all changes were tested in TylerTesting which can be found in Assets > Scenes > TestScenes > TylerTesting