Grass Cube / New Grass Texture
Implemented another tileable grass texture that's darker than the one previously implemented. This allows for the removal of borders between tiles since it's very visually distinct from other tiles around it.
A new model was added - a 3d cube. This cube is what the grass textures will be applied to.