wrong tile index bugfix
There was a bug with grids where they tiles would be the wrong index and it causes a lot of bugs with highlight and pathing and stuff, should be fixed now. you can test this in any scene that has a grid by making sure that the grid index of a tile is correct with its position in the grid.