Objects not childing to the tile bug fix
for the player, melee enemy, and obstacle prefab i shrunk the hitbox and turned them into triggers, this fixed the issue of them childing to the wrong tile. You should be able to test this on any scene with a working grid but I specifically tested it in the MartaTesting scene.